ApproximateGPs API

ApproximateGPs.Centered — Type

Centered()

Used in conjunction with SparseVariationalApproximation. States that the q field of SparseVariationalApproximation is to be interpreted directly as the approximate posterior over the pseudo-points.

This is also known as the "unwhitened" parametrization [1].

See also NonCentered.

[1] - https://en.wikipedia.org/wiki/Whitening_transformation

ApproximateGPs.NonCentered — Type

NonCentered()

Used in conjunction with SparseVariationalApproximation. States that the q field of SparseVariationalApproximation is to be interpreted as the approximate posterior over cholesky(cov(u)).L \ (u - mean(u)), where u are the pseudo-points.

This is also known as the "whitened" parametrization [1].

See also Centered.

[1] - https://en.wikipedia.org/wiki/Whitening_transformation

ApproximateGPs.SparseVariationalApproximation — Method

SparseVariationalApproximation(fz::FiniteGP, q::AbstractMvNormal)

Packages the prior over the pseudo-points fz, and the approximate posterior at the pseudo-points, which is mean(fz) + cholesky(cov(fz)).L * ε, ε ∼ q.

Shorthand for

SparseVariationalApproximation(NonCentered(), fz, q)

ApproximateGPs.SparseVariationalApproximation — Method

SparseVariationalApproximation(
    ::Parametrization, fz::FiniteGP, q::AbstractMvNormal
) where {Parametrization}

Produce a SparseVariationalApproximation{Parametrization}, which packages the prior over the pseudo-points, fz, and the approximate posterior at the pseudo-points, q, together into a single object.

The Parametrization determines the precise manner in which q and fz are interpreted. Existing parametrizations include Centered and NonCentered.

AbstractGPs.elbo — Method

elbo(
    sva::SparseVariationalApproximation,
    fx::FiniteGP,
    y::AbstractVector{<:Real};
    num_data=length(y),
    quadrature=DefaultQuadrature(),
)

Compute the Evidence Lower BOund from [1] for the process f = fx.f == svgp.fz.f where y are observations of fx, pseudo-inputs are given by z = svgp.fz.x and q(u) is a variational distribution over inducing points u = f(z).

quadrature selects which method is used to calculate the expected loglikelihood in the ELBO. The options are: DefaultQuadrature(), Analytic(), GaussHermite() and MonteCarlo(). For likelihoods with an analytic solution, DefaultQuadrature() uses this exact solution. If there is no such solution, DefaultQuadrature() either uses GaussHermite() or MonteCarlo(), depending on the likelihood.

N.B. the likelihood is assumed to be Gaussian with observation noise fx.Σy. Further, fx.Σy must be isotropic - i.e. fx.Σy = α * I.

[1] - Hensman, James, Alexander Matthews, and Zoubin Ghahramani. "Scalable variational Gaussian process classification." Artificial Intelligence and Statistics. PMLR, 2015.

AbstractGPs.elbo — Method

elbo(
    sva::SparseVariationalApproximation,
    lfx::LatentFiniteGP,
    y::AbstractVector;
    num_data=length(y),
    quadrature=DefaultQuadrature(),
)

Compute the ELBO for a LatentGP with a possibly non-conjugate likelihood.

AbstractGPs.posterior — Method

posterior(la::LaplaceApproximation, lfx::LatentFiniteGP, ys)

Construct a Gaussian approximation q(f) to the posterior p(f | y) using the Laplace approximation. Solves for a mode of the posterior using Newton's method.

AbstractGPs.posterior — Method

posterior(sva::SparseVariationalApproximation{Centered})

Compute the approximate posterior [1] over the process f = sva.fz.f, given inducing inputs z = sva.fz.x and a variational distribution over inducing points sva.q (which represents $q(u)$ where u = f(z)). The approximate posterior at test points $x^*$ where $f^* = f(x^*)$ is then given by:

\[q(f^*) = \int p(f | u) q(u) du\]

which can be found in closed form.

[1] - Hensman, James, Alexander Matthews, and Zoubin Ghahramani. "Scalable variational Gaussian process classification." Artificial Intelligence and Statistics. PMLR, 2015.

AbstractGPs.posterior — Method

posterior(sva::SparseVariationalApproximation{NonCentered})

Compute the approximate posterior [1] over the process f = sva.fz.f, given inducing inputs z = sva.fz.x and a variational distribution over inducing points sva.q (which represents $q(ε)$ where ε = cholesky(cov(fz)).L \ (f(z) - mean(f(z)))). The approximate posterior at test points $x^*$ where $f^* = f(x^*)$ is then given by:

\[q(f^*) = \int p(f | ε) q(ε) du\]

which can be found in closed form.

[1] - Hensman, James, Alexander Matthews, and Zoubin Ghahramani. "Scalable variational Gaussian process classification." Artificial Intelligence and Statistics. PMLR, 2015.

ApproximateGPs.approx_lml — Method

approx_lml(la::LaplaceApproximation, lfx::LatentFiniteGP, ys)

Compute an approximation to the log of the marginal likelihood (also known as "evidence"), which can be used to optimise the hyperparameters of lfx.

This should become part of the AbstractGPs API (see JuliaGaussianProcesses/AbstractGPs.jl#221).

ApproximateGPs.build_laplace_objective — Method

build_laplace_objective(build_latent_gp, xs, ys; kwargs...)

Construct a closure that computes the minimisation objective for optimising hyperparameters of the latent GP in the Laplace approximation. The returned closure passes its arguments to build_latent_gp, which must return the LatentGP prior.

Keyword arguments

newton_warmstart=true: (default) begin Newton optimisation at the mode of the previous call to the objective

newton_callback: called as newton_callback(fnew, cache) after each Newton step

newton_maxiter=100: maximum number of Newton steps.

ApproximateGPs.expected_loglik — Method

expected_loglik(quadrature::QuadratureMethod, y::AbstractVector, q_f::AbstractVector{<:Normal}, lik)

This function computes the expected log likelihood:

\[ ∫ q(f) log p(y | f) df\]

where p(y | f) is the process likelihood. This is described by lik, which should be a callable that takes f as input and returns a Distribution over y that supports loglikelihood(lik(f), y).

q(f) is an approximation to the latent function values f given by:

\[ q(f) = ∫ p(f | u) q(u) du\]

where q(u) is the variational distribution over inducing points (see elbo). The marginal distributions of q(f) are given by q_f.

quadrature determines which method is used to calculate the expected log likelihood - see elbo for more details.

Extended help

q(f) is assumed to be an MvNormal distribution and p(y | f) is assumed to have independent marginals such that only the marginals of q(f) are required.

ApproximateGPs.expected_loglik — Method

expected_loglik(::DefaultQuadrature, y::AbstractVector, q_f::AbstractVector{<:Normal}, lik)

The expected log likelihood. Defaults to a closed form solution if it exists, otherwise defaults to Gauss-Hermite quadrature.

ApproximateGPs.laplace_f_and_lml — Method

laplace_f_and_lml(lfx::LatentFiniteGP, ys; newton_kwargs...)

Compute a mode of the posterior and the Laplace approximation to the log marginal likelihood.

ApproximateGPs.laplace_f_cov — Method

laplace_f_cov(cache)

Compute the covariance of q(f) from the results of the training computation that are stored in a LaplaceCache.

ApproximateGPs.laplace_lml — Method

laplace_lml(lfx::LatentFiniteGP, ys; newton_kwargs...)

Compute the Laplace approximation to the log marginal likelihood.

ApproximateGPs.laplace_steps — Method

laplace_steps(lfx::LatentFiniteGP, ys; newton_kwargs...)

For demonstration purposes: returns an array of all the intermediate approximations of each Newton step.

If you are only interested in the actual posterior, use posterior(::LaplaceApproximation, ....

TODO figure out how to get the @ref to work to point to the LaplaceApproximation-specific posterior docstring...

ApproximateGPs.loglik_and_derivs — Method

loglik_and_derivs(dist_y_given_f, ys, f)

dist_y_given_f must be a scalar function from a Real to a Distribution object. ys and f are vectors of observations and latent function values, respectively.

ApproximateGPs.newton_inner_loop — Method

newton_inner_loop(dist_y_given_f, ys, K; f_init, maxiter, callback=nothing)

Find a mode of p(f | y) using Newton's method.