Sparse Variational Approximation

ApproximateGPs.SparseVariationalApproximationModule.Centered — Type

Centered()

Used in conjunction with SparseVariationalApproximation. States that the q field of SparseVariationalApproximation is to be interpreted directly as the approximate posterior over the pseudo-points.

This is also known as the "unwhitened" parametrization [1].

See also NonCentered.

[1] - https://en.wikipedia.org/wiki/Whitening_transformation

ApproximateGPs.SparseVariationalApproximationModule.NonCentered — Type

NonCentered()

Used in conjunction with SparseVariationalApproximation. States that the q field of SparseVariationalApproximation is to be interpreted as the approximate posterior over cholesky(cov(u)).L \ (u - mean(u)), where u are the pseudo-points.

This is also known as the "whitened" parametrization [1].

See also Centered.

[1] - https://en.wikipedia.org/wiki/Whitening_transformation

ApproximateGPs.SparseVariationalApproximationModule.SparseVariationalApproximation — Method

SparseVariationalApproximation(fz::FiniteGP, q::AbstractMvNormal)

Packages the prior over the pseudo-points fz, and the approximate posterior at the pseudo-points, which is mean(fz) + cholesky(cov(fz)).L * ε, ε ∼ q.

Shorthand for

SparseVariationalApproximation(NonCentered(), fz, q)

ApproximateGPs.SparseVariationalApproximationModule.SparseVariationalApproximation — Method

SparseVariationalApproximation(
    ::Parametrization, fz::FiniteGP, q::AbstractMvNormal
) where {Parametrization}

Produce a SparseVariationalApproximation{Parametrization}, which packages the prior over the pseudo-points, fz, and the approximate posterior at the pseudo-points, q, together into a single object.

The Parametrization determines the precise manner in which q and fz are interpreted. Existing parametrizations include Centered and NonCentered.